Add the '--enable-print-device-tree' argument
I'm trying to debug some device tree problems while booting Linux and
figured it would be really nice to have access to the device tree while
trying to debug these problems.  I think this might be useful for lots
of people, so I went ahead and cleaned up the code enough that it should
actaully work in most cases.