Allow platforms to limit the physical memory size
The DMA controllor on the VC707 only supports 32-bit physical addresses.
Since we don't have an IOMMU or DMA regions in our Linux port, we need a
mechanism for limiting the amount of physical memory that Linux sees.
Since we map the memory to 2GB physical, we can only address 2GB of
DRAM.

Like everything in platform/, this is a hack.  This one we won't replace
with a DTS parser, we'll actually entirely remove it from BBL and
replace it with support for small DMA regions in our Linux port.