Allow 32-bit HTIF because it works on qemu-system-riscv32
To enable testing of rv32 riscv-linux on qemu-system-riscv32
it would be good if we had working HTIF in bbl. This change
has been tested using riscv-probe in spike and qemu:

- https://github.com/michaeljclark/riscv-probe

While there are compiler issues that could re-order 64-bit
loads and stores, and there is no official HTIF over 32-bit
protocol, or 64-bit HTIF with ordered loads/stores and fences;
however, given that it works as it is, it should be safe to
re-enable 32-bit support with a caveat.

Signed-off: mjc@sifive.com