Replace the --enable-32bit option by --with-arch
Get rid of the --enable-32bit option and switch to use --with-arch,
which is more standard because it matches the GCC build.

If --with-arch is not specified, it defaults to whatever the
compiler's default is.

The --with-abi is not necessary for this project.
Unconditionally compile it with a no-float ABI.