Fix trap handler when DRAM starts above 32-bits
The trap handler used the lower half of the PC to compute an offset in
the trap_vector. This is fine as long asthe top half of the PC address
is all zero, but when BBL is mapped at or above 0x100000000 the computed
address is wrong. This in turn leads us to take another trap when we use
the incorrect offset, locking us in an infinite trap loop.

We have to also look at the top half of the PC address, and re-add it
when we've figured out the trap_vector offset.

Sponsored by:   Axiado