machine: manually perform assembler relaxation
This is an equivalent rewrite of the existing code.  When building with
gas, the `bltu` would implicitly get relaxed to the `bgeu` + `j`.  This
relaxation is required as the `init_other_hart` is not guaranteed to be
addressable in 12-bits.  When building with the LLVM IAS instead of gas
we fail to link as the branch is not relaxed.  This change enables LLVM
to build and link this code with the LLVM IAS and lld.