machine: correct some additional cases of UB
Use of asm aliased register variables in local scope can only be used
for extended assembly parameters.  This changes the few instances of
this in the floating point emulation to use the GNU extended assembly
syntax to access the `tp` register.  This ensures that we do not rely on
undefined behaviour.  This was uncovered when building the Proxy kernel
with clang and LLVM.