Add ENOSYS stub for syscall 278, getrandom
Binaries built with a recent LLVM and current riscv-gnu-toolchain
and targeting riscv64-unknown-linux-gnu fail to execute under Spike+pk
with this output:
'''
$ /scratch/projects/riscv-isa-sim//build/spike \
    --varch=vlen:128,elen:64 --isa=rv64gcv_zvkned \
    /rivos/riscv-pk/riscv64-unknown-linux-gnu/bin/pk aes-cbc-test
bbl loader
bad syscall #278!
'''

Syscall 278 is getrandom which is invoked prior to reaching main().
Providing a syscall stub to return ENOSYS is enough to please
the caller. This has the additional benefit of not requiring
to implement getrandom() in a portable way.

'''
$ /scratch/projects/riscv-isa-sim//build/spike \
    --varch=vlen:128,elen:64 --isa=rv64gcv_zvkned \
    /rivos/riscv-pk/riscv64-unknown-linux-gnu/bin/pk aes-cbc-test
bbl loader
vlen: 128 bits
--- Running 'CBCKeySbox256' test suite...
...
'''

Signed-off-by: Eric Gouriou <ego@rivosinc.com>